+++ tests/array_interoperability_test.py
@@ -195,13 +195,14 @@ class DLPackTest(jtu.JaxTestCase):
     shape=all_shapes,
     dtype=numpy_dtypes,
     copy=[False, True],
+    device_transfer=[False, True],
   )
-  def testNumpyToJax(self, shape, dtype, copy):
+  def testNumpyToJax(self, shape, dtype, copy, device_transfer):
     rng = jtu.rand_default(self.rng())
     x_np = rng(shape, dtype)
-    device = jax.devices()[0]
+    device = jax.devices()[0] if device_transfer else None
     _from_dlpack = lambda: jnp.from_dlpack(x_np, device=device, copy=copy)
-    if jax.default_backend() == 'gpu' and not copy:
+    if device_transfer and not copy:
       self.assertRaisesRegex(
         ValueError,
         r"Specified .* which requires a copy",